Extract common CSV report building code.
Changes the name of the "all-editions" report as
a byproduct, however this name wasn't very clear
and the report is almost or entirely unused.
